Spot Light

+++ src/PointLight.cpp
@@ -21,9 +21,9 @@ PointLight::PointLight(GLfloat red, GLfloat green, GLfloat blue,
 
 
 
-void PointLight::UseLight(GLfloat ambientIntensityLocation, GLfloat ambientColorLocation,
-        GLfloat difuseIntensityLocation, GLfloat positionLocation,
-        GLfloat constantLocation, GLfloat linearLocation, GLfloat exponentLocation)
+void PointLight::UseLight(GLuint ambientIntensityLocation, GLuint ambientColorLocation,
+        GLuint difuseIntensityLocation, GLuint positionLocation,
+        GLuint constantLocation, GLuint linearLocation, GLuint exponentLocation)
 {
     glUniform3f(ambientColorLocation, colour.x, colour.y, colour.z);
     glUniform1f(ambientIntensityLocation, ambientIntensity);
